Adjust tile bounds so isochrones remain centered (#705)
* Add 10 minutes to max_minutes (same as what is done in the isochrone service).

* Add a method to nudge the tile bounds. This can be useful to keep a center location
centered within a tile/grid as different sized bounds are created.

* Add a constant to convert MPH to meters per second.

* Add code to shift the bounding box such that the center location lies within
the center of a tile/grid. This keeps the "grid" from shifting with different
isochrone calls centered at the same location but with different max times.

* Fix typo

* Add seconds to minutes constants

* Add ExpandForward and ExpandReverse methods to simplify transition logic.
Optimize the grid size to create an isotile roughly 500 rows - but round
grid size to nearest 0.001 increment so it doesn't rapidly fluctuate.
Return if the cost exceeds the max_time by a factor of 4 (in case the
secs threshold is not triggered).

* Use smaller grid size for multimodal (since the max distance is larger but the
generation time is faster). Skip over highly penalized edges within multimodal
isochrones.

* Default generalization for contours is a special value that indicates the method
should compute an optimal generalization factor (currenly 1/8 the grid size).

* Move back to 1/8 grid size for default generalization factor.

* Use max float as the trigger for optimal generalization tolerance.