Add compressed speed profiles (#1410)
* Initial commit of predicted speeds additions to graph tiles. Added a PredictedSpeeds class
to get a speed given a time bucket within the week. Change the GetSpeeds method to add an edgeId
and a seconds within the week argument to allow lookup within the bucketized speeds. Add
information to GraphTileHeader to enable an offset and count for PredictedSpeed records.

* Change PredictedSpeed to have 2 methods to set the indexes and speed profile pointers
within the GraphTile. Change so it is not a pointer within GraphTile.

* Speed profiles need to be signed int16

* Add UpdatePredictedSpeeds to GraphTileBuilder.

* Update formatting.

* Updates the the DTC III method.

* Add normalization to the DTC III methods. Avoid using Boost base64 decoding - it gives a result size = 402
bytes rather than the expected 401 bytes.

* Updated DTC III - no efficiency simplifications have been made. Want to validate and get the right answers,
then optimize.

* Update DTC III method to use speed normalizations

* Updates to working initial version - uses Boost for base64 decoding, properly converts
big endian values to little endian values, uses PredictedSpeeds class to compare
decoded values against expected values.

* Add PredictedSpeed information

* Update valhalla_add_predicted_speeds to add the compressed profiles. Formatting updates

* SOme cleanup. No need to iterate through nodes, just iterate all directed edges.

* Fix formatting

* Remove sqrt from constexpr to fix OSX builds