Isotile optimization (#1506)
* Add resample_polyline. This is a faster implementation than sample_spherical_polyline since it
only does linear interpolation in lat,lon space. The accuracy is less, but over short distances
this should be OK> Its primary use is in generating isotiles for the isochrone service which
resample shape of road edges - generally over short distances between vertices. Since isochrones
skip ferries and transit edges (which are more typically the longerst straight edges) this
hopefully will not cause any issues.

* Remove extraneuous marking of edges using the shape points of the edge (this likely
should have been removed when the resampled polyline code was added). Change to use
the new resample_polyline method (to avoid much of the trigonometric methods used in
resmaple_spherical_polyline). Rather than using the Intersect method (which uses a
Bresenham algorithm to walk tiles along the segment), use a simple method to mark
any tiles within a bounding box of the polyline segment. Since resampling occurs this
approximation should not cause too much inaccuracy.

* Use the double precision Distance method in PointLL. Trying to use a float variant
gave incorrect answers and created a large number of vertices in many of the
resampled polylines. This version consistently gives n < (length/resolution + 2)

* The Intersect method (using Bresenham algorithm) creates a better visual result but is slower. This change
avoids calling the Intersect method if the segments endpoints are in the same tile or are neighbors directly
east, west, north, or south of each other. In the latter case only the segment endpoints are touched. This
change improves performance over always using the Intersect method, but gives a cleaner visualization than
by using a simpler method such as the marking tiles intersecting the bounding box around the segment.

* Compute delta using the size of the resampled shape vector rather than the ratio of
lengths.

* Use similar method to determine intersecting tiles within the "short edge" code as within
the resampled shape code. Also extend the threshold to be 1.25 * the shape_interval.