Add a has_traffic flag to TransitionCost and TransitionCostReverse.  (#1553)
* Add a has_traffic flag to TransitionCost and TransitionCostReverse. This defaults to
false. For time dependent rotues where traffic is available at an intersection we want
to lower the influence of transition costs (especially when computing estimated time
to traverse the intersection). This change allows future use within all costing models
that need to be influenced by traffic, but for now only time dependent, depart-at
routes are set to use this flag.

* Had to override this TransitionCost.

* And the reverse

* Add has_traffic flags to calls to transition costing within timedep_reverse and isochrone

* format.

* Group speeds together in a JSON object.