Add durations to allow trace_route to compute elapsed times using the trace (#1679)
* Add durations for between each pair of shape points. This is one way to
get elapsed times (timestamps) set at each shape point for trace_route
or trace_attributes.

* Add use_timestamps option for trace_route (perhaps trace_attributes?). If timestamps are
present for the trace, if this option is true then the timestamp values are used to compute
elapsed time along the path. Otherwise costing methods are used to compute elapsed time. This
commit adds the option and use of the option for RouteMatch (edge walking).

* Add use_timestamps to map matching FormPath. If this is true the timestamps on
the trace will be used to compute elapsed time (otherwise costing is used).

* Formatting

* Add error checking for durations and use_timestamps

* Update ChangeLog

* Update map-matching doc to include durations and use_timestamps.

* Add begin_time as an option for populating timestamps for a trace input using encoded polyline.

* Update api-reference.md

* To make this more complete, a begin_time option has been added. This is the starting timestamp of the
input trace. durations are added to that to create real timestamps for each trace point. This allows
timestamps to be supplied when encoded polylines are used to specify the trace points

* Update CHANGELOG.md