valhalla_build_tiles - Allow to run as sub-stages (#1705)
* Add names for the temporary bin files to the functions (they are passed in
externally so they can be cleaned up/deleted when done).

* Pass in the temporary bin filenames so they can be cleaned up (deleted) externally
when done.

* Add a start and end stage and allow the pipeline to be broken into sub-stages. Still need
to add command line arguments to control this, but the plumbing is in place. Added conditionals
to the running of each stage of the pipeline. Also added a cleanup and centralized the list
of temporary files (*.bin) so they can be deleted during the cleanup stage. These are placed
in the tile_dir specified in the config file so they cannot be mixed with temporary files
from another sub-stage/config.

* Fix formatting

* Fix tests and clean up temporary files build during tests.

* Add command line arguments for start and end stage to run in the build tiles pipeline.

* Fix formatting

* Test updating filename

* Remove unused boost headers

* If error with file in the sequence class, add sequence to exception so we know where it comes from.

* Fix formatting

* Try adding files to local dir rather than test data directory.

* Move temp files back to regular directory.

* remove unused tiledir variable

* Keep sorting of complex restrictions scoped so the file handles don't remain open.

* Make kInvalid equal -1 (change enum to be int8). Move some constant strings to anonymous namespace

* Add changes for valhalla_build_tiles