Fix bug in TrimShape  (#1935)
* Fix bug in TrimShape where if start_vertex is not valid it never trims the start of
the shape but accumulates the along distance and never resets it. This leads to
erasing most of the shape when clipping after the end point. The fix is to not trim
start shape if the start_vertex is not valid since it can never actually erase any of the
shape in this case. Similarly for trimming past the end shape.

* Update change log