Increases kMaxIterationsWithoutConvergence in forward astar
We have a set of problematic routes on the Autobahn which ends with
a rest-stop on the opposite side of travel direction, thereby
requiring overshooting the destination and coming back in the
other direction.

The bidirectional a-star routes fine but the forward timedependant requires
more iterations without convergence for these routes.