Polyline Resampling Addition and Bug Fix (#2239)
* Add uniform_resample_spherical_polyline to resample to a specified
number of vertices (sampling interval is based on the polyline length
and the number of vertices). Fix a bug where duplicate lat,lng locations
in the polyline could cause sampling to be cut short due to NaN in
the distance computation (PointLL::Distance protects against this). Added
a test to verify that duplicate vertices do not shortcut the sampling.

* Update CHANGELOG

* Formatting update

Co-authored-by: Kevin Kreiser <kevinkreiser@gmail.com>