Don't allow more than one dot when parsing numbers in Lua. (#2371)
* Better unit parsing for heights in tags.

Added a unit test framework for checking that parsed heights are giving the results that we expect. Used this to rebuild the height parsing routing in Lua so that it passes all the tests.

* Belt and braces approach to parsing numbers.

The problem is that the `numeric_prefix` function would accept a number with multiple dots in, e.g: `"3..35"`. Unfortunately, calling `tonumber()` on this would return `nil`, and then Lua would barf when trying to do arithmetic on a `nil` value.

* Fix bug and provide better error message when Lua barfs.

There was a bug when `lua_pcall` returned something other than success. This caused a log message to be printed, but the function then continued into pulling the results off the Lua stack. If these results didn't exist, it was possible to sometimes get SIGSEGV.

Since the Lua code is part of Valhalla, it makes sense that a bad response from Lua should be treated as a logic error and exceptional condition. However, this wouldn't give a lot of information for debugging, so I've added the OSM ID as a parameter to provide better debuggability.

* Clang formatting "fixes".

Co-authored-by: Matt Amos <zerebubuth@gmail.com>