Support for Mid-Edge Uturns when Map Matching (#2380)
At the moment, the serializer that turns a map match into a route will not allow you to make a uturn in the middle of an edge in the graph. This leads to much longer routes than make sense especially for pedestrians and cyclists.

This PR allows uturns in the middle of an edge and provides the correct guidance for that as well. As wtih routing, the location types (break, via, through and break_through) are used to dictate whether you'd like the matcher to allow a uturn at a given trace point or not. Previously the matcher always allowed making a uturn at a given trace point but heavily penalized it. Now there is no penalty but simply whether or not uturns are allowed based on the location type.

Co-authored-by: Buro Mookerji <buro.mookerji@mapbox.com>