Enhanced turn lane information (#2653)
The turn lane object (intersection.lanes) within a step is enhanced with
new attributes:
* "active": Set to true for lanes which are most appropriate to be in,
considering the current and next maneuver. false otherwise.
* "valid_indication": An optional attribute containing the active lane's
arrow indication in the direction of the maneuver. Eg: if active lane has
"indications: [ left, straight ]" and we're turning left from this lane,
then "valid_indication" will be "left".

The existing attribute "valid", will be set to true for lanes which can
be taken for completing the maneuver, but might not be the best lane to
be in, considering the next maneuver.